Split monolithic file into project structure
