+++ .env.local
@@ -1,0 +1,2 @@
+NEXT_PUBLIC_GA_ID=G-XXXXXXXXXX
+NEXT_PUBLIC_SITE_URL=https://mayanktrivedi.com
