+++ .eslintrc.json
@@ -1,0 +1,8 @@
+{
+  "extends": ["next", "next/core-web-vitals"],
+  "rules": {
+    "react/prop-types": "off",
+    "jsx-a11y/anchor-is-valid": "off"
+  }
+}
+
