+++ README.md
@@ -1,0 +1,43 @@
+# Mayank Trivedi Portfolio
+
+A premium, enterprise-grade portfolio website built with Next.js, React, and Tailwind CSS.
+
+## Features
+
+- 🎨 Elegant, professional design with custom branding
+- 📱 Fully responsive and mobile-optimized
+- ♿ WCAG 2.1 AA compliant for accessibility
+- 🚀 Optimized for performance (target Lighthouse score: 100)
+- 🎭 Interactive 3D avatar with multiple modes
+- 🗺️ Curiosity Passport tracking user journey
+- 📊 Timeline slider with scroll synchronization
+- 🎯 Smart CTAs based on user behavior
+- 🥚 Professional easter eggs
+- 🔊 Optional audio features
+- 📱 PWA support with offline capabilities
+- 🖨️ Print-friendly layouts
+
+## Getting Started
+
+1. Clone the repository
+2. Install dependencies: `npm install`
+3. Run development server: `npm run dev`
+4. Build for production: `npm run build`
+5. Export static site: `npm run export`
+
+## Deployment
+
+This site is optimized for deployment on Vercel or Netlify.
+
+## Tech Stack
+
+- Next.js 14
+- React 18
+- Tailwind CSS
+- Framer Motion
+- React Three Fiber
+- React Hook Form
+
+## License
+
+© 2024 Mayank Trivedi. All rights reserved.
