+++ components/ContactForm.jsx
@@ -1,4 +1,3 @@
-// components/ContactForm.jsx
 import React, { useState } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
 import { useForm } from 'react-hook-form';
@@ -291,119 +290,3 @@ const ContactForm = ({ initialSubject = '' }) => {
 
 export default ContactForm;
 
-// public/manifest.json
-{
-  "name": "Mayank Trivedi - Visionary Leader",
-  "short_name": "Mayank Trivedi",
-  "description": "CEO, entrepreneur, and craftsman building the future of healthcare technology.",
-  "start_url": "/",
-  "display": "standalone",
-  "background_color": "#FAFAF9",
-  "theme_color": "#0A2540",
-  "orientation": "portrait",
-  "icons": [
-    {
-      "src": "/icon-192.png",
-      "sizes": "192x192",
-      "type": "image/png"
-    },
-    {
-      "src": "/icon-512.png",
-      "sizes": "512x512",
-      "type": "image/png"
-    }
-  ]
-}
-
-// public/robots.txt
-User-agent: *
-Allow: /
-
-Sitemap: https://mayanktrivedi.com/sitemap.xml
-
-// utils/helpers.js
-export const formatDate = (date) => {
-  return new Intl.DateTimeFormat('en-US', {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-  }).format(new Date(date));
-};
-
-export const scrollToTop = () => {
-  window.scrollTo({
-    top: 0,
-    behavior: 'smooth',
-  });
-};
-
-export const debounce = (func, wait) => {
-  let timeout;
-  return function executedFunction(...args) {
-    const later = () => {
-      clearTimeout(timeout);
-      func(...args);
-    };
-    clearTimeout(timeout);
-    timeout = setTimeout(later, wait);
-  };
-};
-
-export const isBrowser = () => typeof window !== 'undefined';
-
-export const getScrollPercent = () => {
-  if (!isBrowser()) return 0;
-  
-  const scrollTop = window.pageYOffset || document.documentElement.scrollTop;
-  const scrollHeight = document.documentElement.scrollHeight - document.documentElement.clientHeight;
-  return (scrollTop / scrollHeight) * 100;
-};
-
-// .env.local (example)
-NEXT_PUBLIC_GA_ID=G-XXXXXXXXXX
-NEXT_PUBLIC_SITE_URL=https://mayanktrivedi.com
-
-// README.md
-# Mayank Trivedi Portfolio
-
-A premium, enterprise-grade portfolio website built with Next.js, React, and Tailwind CSS.
-
-## Features
-
-- 🎨 Elegant, professional design with custom branding
-- 📱 Fully responsive and mobile-optimized
-- ♿ WCAG 2.1 AA compliant for accessibility
-- 🚀 Optimized for performance (target Lighthouse score: 100)
-- 🎭 Interactive 3D avatar with multiple modes
-- 🗺️ Curiosity Passport tracking user journey
-- 📊 Timeline slider with scroll synchronization
-- 🎯 Smart CTAs based on user behavior
-- 🥚 Professional easter eggs
-- 🔊 Optional audio features
-- 📱 PWA support with offline capabilities
-- 🖨️ Print-friendly layouts
-
-## Getting Started
-
-1. Clone the repository
-2. Install dependencies: `npm install`
-3. Run development server: `npm run dev`
-4. Build for production: `npm run build`
-5. Export static site: `npm run export`
-
-## Deployment
-
-This site is optimized for deployment on Vercel or Netlify.
-
-## Tech Stack
-
-- Next.js 14
-- React 18
-- Tailwind CSS
-- Framer Motion
-- React Three Fiber
-- React Hook Form
-
-## License
-
-© 2024 Mayank Trivedi. All rights reserved.
