+++ package.json
@@ -1,0 +1,37 @@
+{
+  "name": "mayank-trivedi-portfolio",
+  "version": "1.0.0",
+  "private": true,
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "export": "next export",
+    "lint": "next lint",
+    "analyze": "ANALYZE=true next build"
+  },
+  "dependencies": {
+    "next": "^14.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "@react-three/fiber": "^8.15.0",
+    "@react-three/drei": "^9.88.0",
+    "three": "^0.158.0",
+    "framer-motion": "^10.16.0",
+    "react-intersection-observer": "^9.5.0",
+    "@headlessui/react": "^1.7.0",
+    "react-hook-form": "^7.47.0",
+    "clsx": "^2.0.0",
+    "date-fns": "^2.30.0",
+    "react-icons": "^4.11.0"
+  },
+  "devDependencies": {
+    "autoprefixer": "^10.4.16",
+    "postcss": "^8.4.31",
+    "tailwindcss": "^3.3.5",
+    "eslint": "^8.53.0",
+    "eslint-config-next": "^14.0.0",
+    "@next/bundle-analyzer": "^14.0.0"
+  }
+}
+
