+++ pages/index.js
@@ -1,70 +1,3 @@
-// pages/_app.js
-import '@/styles/globals.css';
-import { PassportProvider } from '@/context/PassportContext';
-import { useEffect } from 'react';
-import { useRouter } from 'next/router';
-import { trackEvent } from '@/utils/analytics';
-
-function MyApp({ Component, pageProps }) {
-  const router = useRouter();
-
-  useEffect(() => {
-    const handleRouteChange = (url) => {
-      trackEvent('page_view', { page_path: url });
-    };
-
-    router.events.on('routeChangeComplete', handleRouteChange);
-    return () => {
-      router.events.off('routeChangeComplete', handleRouteChange);
-    };
-  }, [router.events]);
-
-  return (
-    <PassportProvider>
-      <Component {...pageProps} />
-    </PassportProvider>
-  );
-}
-
-export default MyApp;
-
-// pages/_document.js
-import { Html, Head, Main, NextScript } from 'next/document';
-
-export default function Document() {
-  return (
-    <Html lang="en">
-      <Head>
-        <meta charSet="utf-8" />
-        <link rel="icon" href="/favicon.ico" />
-        <link rel="apple-touch-icon" href="/apple-touch-icon.png" />
-        <meta name="theme-color" content="#0A2540" />
-        
-        {/* Google Analytics */}
-        <script
-          async
-          src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GA_ID}`}
-        />
-        <script
-          dangerouslySetInnerHTML={{
-            __html: `
-              window.dataLayer = window.dataLayer || [];
-              function gtag(){dataLayer.push(arguments);}
-              gtag('js', new Date());
-              gtag('config', '${process.env.NEXT_PUBLIC_GA_ID}');
-            `,
-          }}
-        />
-      </Head>
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  );
-}
-
-// pages/index.js
 import React from 'react';
 import Layout from '@/components/Layout';
 import Hero from '@/components/Hero';
