+++ pages/leadership.js
@@ -1,4 +1,3 @@
-// pages/leadership.js
 import React, { useEffect, useState } from 'react';
 import Layout from '@/components/Layout';
 import Hero from '@/components/Hero';
