+++ public/manifest.json
@@ -1,0 +1,23 @@
+{
+  "name": "Mayank Trivedi - Visionary Leader",
+  "short_name": "Mayank Trivedi",
+  "description": "CEO, entrepreneur, and craftsman building the future of healthcare technology.",
+  "start_url": "/",
+  "display": "standalone",
+  "background_color": "#FAFAF9",
+  "theme_color": "#0A2540",
+  "orientation": "portrait",
+  "icons": [
+    {
+      "src": "/icon-192.png",
+      "sizes": "192x192",
+      "type": "image/png"
+    },
+    {
+      "src": "/icon-512.png",
+      "sizes": "512x512",
+      "type": "image/png"
+    }
+  ]
+}
+
