+++ public/robots.txt
@@ -1,0 +1,4 @@
+User-agent: *
+Allow: /
+
+Sitemap: https://mayanktrivedi.com/sitemap.xml
