+++ styles/globals.css
@@ -1,4 +1,3 @@
-// styles/globals.css
 @import url('https://fonts.googleapis.com/css2?family=Lato:wght@300;400;700;900&family=EB+Garamond:wght@400;500;600&display=swap');
 @tailwind base;
 @tailwind components;
