+++ utils/seo.js
@@ -1,0 +1,24 @@
+export const defaultSEO = {
+  title: 'Mayank Trivedi - Visionary Leader in Healthcare Innovation',
+  description: 'CEO, entrepreneur, and craftsman building the future of healthcare technology with heart and hands.',
+  openGraph: {
+    type: 'website',
+    locale: 'en_US',
+    url: 'https://mayanktrivedi.com/',
+    site_name: 'Mayank Trivedi Portfolio',
+    images: [
+      {
+        url: '/images/og-image.png',
+        width: 1200,
+        height: 630,
+        alt: 'Mayank Trivedi',
+      },
+    ],
+  },
+  twitter: {
+    handle: '@mayanktrivedi',
+    site: '@mayanktrivedi',
+    cardType: 'summary_large_image',
+  },
+};
+
